more optimizations fixed 3rd sub pickup

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/testing/IntakeTesting.java
@@ -15,7 +15,7 @@ public class IntakeTesting extends LinearOpMode {
   NewRobot robot;
 
   public static double rotatePos = Intake.INTAKE_FLAT;
-  public static int hSlidePos = HorizontalSlides.TRANSFER_POS;
+  public static double hSlidePos = HorizontalSlides.TRANSFER_POS;
 
   @Override
   public void runOpMode() throws InterruptedException {
